Landing page UX update
A text in the bottom where added for phones saying to click portfolio with animation up and down making it stick out.

More over font will change a bit for going up the size to ipad. Problem is the placement of the text for the margin-top 12% which changes way to much between touch screens. Would not use again and fix.